refactor(Name) : Name , Names common 패키지에서 player 패키지로 이동
[관련 피드백 링크](https://github.com/woowacourse/java-blackjack/pull/652#discussion_r1522183106)
