+++ src/main/java/blackjack/domain/player/Dealer.java
@@ -2,7 +2,6 @@ package blackjack.domain.player;
 
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
-import blackjack.domain.common.Name;
 import blackjack.domain.result.ResultStatus;
 
 public class Dealer extends Participant {
