+++ src/main/java/blackjack/domain/player/Participant.java
@@ -1,7 +1,6 @@
 package blackjack.domain.player;
 
 import blackjack.domain.card.Card;
-import blackjack.domain.common.Name;
 import blackjack.domain.card.Cards;
 
 import java.util.List;
