+++ src/test/java/blackjack/domain/common/NameTest.java
@@ -2,7 +2,7 @@ package blackjack.domain.common;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import blackjack.domain.common.Name;
+import blackjack.domain.player.Name;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
