+++ src/test/java/blackjack/domain/result/GamePlayerResultTest.java
@@ -2,9 +2,7 @@ package blackjack.domain.result;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import blackjack.domain.common.Name;
-import blackjack.domain.result.GamePlayerResult;
-import blackjack.domain.result.ResultStatus;
+import blackjack.domain.player.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
